+++ typed-protocols/src/Network/TypedProtocol/Proofs.hs
@@ -39,6 +39,7 @@ import           Data.Function ((&))
 import           Data.Kind (Type)
 import           Data.Singletons
 import           Network.TypedProtocol.Core
+import           Network.TypedProtocol.Lemmas
 import           Network.TypedProtocol.Peer
 
 -- $about
